pci: Rework pci_bar_is_valid()
Currently pci_bar_is_valid() returns the value stored
in a BAR. That is wrong, because the function fails if
the returned value is zero. But that is wrong, because:
  - an address in PIO address space may be zero;
  - an upper part of 64-bit BAR may be zero;

The right way of validating a BAR is checking its size -
for implemented BARs the size is non-zero.

It might look strange, but the function checks not just the
queried BAR, but sanity of all six BARs of the device.

If we checked only requested BAR the function could have
succeeded with inconsistent BAR sizes combinations. I.e.
imagine a corrupted BARs reported by a device:
  BAR 0 - 64 bit lower
  BAR 1 - 64 bit lower
  BAR 2 - 64 bit upper
  ...
If the function is asked about BAR 1 it would succeed, even
though in fact the layout is ambiguous.

As result, the function is expected to be called only with
valid BAR numbers and BARs layout should be consistent.
Otherwise, an assert is raised.

The BAR number parameter bar_num is only allowed to refer
a 32-bit BAR or lower part of a 64-bit BAR, as it does not
make sense checking validity of high part of a 64-bit BAR.
Thus, the caller is expected to know the device BARs layout.

Cc: Thomas Huth <thuth@redhat.com>
Cc: Andrew Jones <drjones@redhat.com>
Signed-off-by: Alexander Gordeev <agordeev@redhat.com>